refactor(core): migrar Double a BigDecimal para precisión volumétrica y financiera
Cambios realizados:
- Refactorización de Value Objects y Entidades de dominio para asegurar precisión decimal.
- Implementación de comparaciones mediante .compareTo() y operaciones aritméticas inmutables.
- Actualización de lógica de validación de capacidad en Locations para prevenir errores de redondeo en el cálculo de peso y volumen.
- Ajuste de comandos y excepciones de dominio para mantener consistencia en la capa de aplicación.

Motivación: Evitar la pérdida de precisión inherente al punto flotante binario en operaciones de misión crítica, cumpliendo con los estándares de auditoría de inventarios industriales.

+++ src/main/java/com/juanbenevento/wms/inventory/domain/event/StockShippedEvent.java
@@ -1,9 +1,10 @@
 package com.juanbenevento.wms.inventory.domain.event;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record StockShippedEvent(
         String sku,
-        Double quantity,
+        BigDecimal quantity,
         String locationCode,
         String username,
         LocalDateTime occurredAt
